AtomicInt32: select atomic functions by compiler
When using MSVC, win32 is implied and the Interloked* functions can be
used.
On Linux, BSD and OSX gcc is available and provides its own atomic
builtins.
Note that Apple has switched to clang a few years ago, however clang
attempts to be source-code-compatible with gcc and thus should work too.